src/requirements.txt
boto3==1.28.19
botocore==1.31.19
cloudevents==1.9.0
kafka-python==2.0.2